UAV/GPS/GPS_Module/requirements_gps.txt
RPi.GPIO
pyserial
numpy
scikit-learn
requests